requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
numpy>=1.26.0
plotly>=5.20.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pyarrow>=14.0.0
apscheduler>=3.10.0
pytest>=7.4.0
pytest-cov>=4.1.0
